server/requirements.txt
Flask==2.3.2
Flask_Cors==4.0.0
opencv_python==4.9.0.80
opencv_python_headless==4.9.0.80
Pillow==10.0.1
Pillow==10.3.0
